Add topics to conference proposals
