+++ src/lib/topic/types.ts
@@ -1,0 +1,10 @@
+export interface Topic {
+  _id: string
+  _type: 'topic'
+  title: string
+  description?: string
+  color: string
+  slug: {
+    current: string
+  }
+}
